+++ frontend/src/app/pages/orcamento-funcionario/orcamento-funcionario.component.ts
@@ -1,9 +1,9 @@
-import { Component } from "@angular/core";
-import { materialImports } from "../../material-imports";
-import { NavbarComponent } from "../../components/navbar/navbar.component";
-import { MatInputModule } from "@angular/material/input";
-import { FormsModule } from "@angular/forms";
-import { CommonModule, CurrencyPipe } from "@angular/common";
+import { Component, OnInit } from '@angular/core';
+import { materialImports } from '../../material-imports';
+import { MatInputModule } from '@angular/material/input';
+import { FormsModule } from '@angular/forms';
+import { CommonModule, CurrencyPipe } from '@angular/common';
+import { NavbarComponent } from '../../components/navbar/navbar.component';
 
 @Component({
   selector: "app-orcamento-funcionario",
@@ -18,9 +18,35 @@ import { CommonModule, CurrencyPipe } from "@angular/common";
   templateUrl: "./orcamento-funcionario.component.html",
   styleUrl: "./orcamento-funcionario.component.css",
 })
-export class OrcamentoFuncionarioComponent {
-  valorOrcamento: string = "";
+
+export class OrcamentoFuncionarioComponent implements OnInit{
+  
+  valorOrcamento: string = '';
+  solicitacao: any = {};
+
+  ngOnInit(): void {
+    this.solicitacao = {
+      id: 12938,
+      equipamento: 'Teclado DELL KB216 USB',
+      problema: 'Teclas ctrl, alt e del não funcionam, espaço solto e barulhento.',
+      cliente: {
+        nome: 'Joana Joaquina',
+        cpf: '000.000.000-00',
+        email: 'joana.joaquina@gmail.com',
+        telefone: '(00) 00000-0000',
+        endereco: {
+          cep: '00000-000',
+          logradouro: 'R. Dr. Alcides Vieira Arcoverde, 1225',
+          complemento: 'Bloco A',
+          cidade: 'Curitiba',
+          estado: 'Paraná'
+        }
+      }
+    };
+  };
+
   constructor(public currencyPipe: CurrencyPipe) {}
+
   formatarMoeda() {
     let valor = this.valorOrcamento.replace(/\D/g, "");
     let valorFormatado = (Number(valor) / 100).toFixed(2);
@@ -30,6 +56,14 @@ export class OrcamentoFuncionarioComponent {
   }
 
   enviarOrcamento() {
-    console.log("BUM! ENVIADO (mentira, só depois de criar um serviço aqui)");
-  }
+    const dataHoraAtual = new Date();
+
+    const registroOrcamento = {
+      valor: this.valorOrcamento,
+      funcionario: 'Funcionario logado',
+      dataHora: dataHoraAtual.toLocaleString('pt-BR')
+    };
+
+    console.log('BUM! ENVIADO! Orçamento registrado: ', registroOrcamento);
+  };
 }
